backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
peewee==3.16.3
python-dotenv==1.0.0